enable swaggerUI

+++ Mall-Tiny-01/src/main/java/me/levi/mall/tiny/mbg/mapper/PmsBrandMapper.java
@@ -6,7 +6,7 @@ import me.levi.mall.tiny.mbg.model.PmsBrandExample;
 import org.apache.ibatis.annotations.Param;
 
 public interface PmsBrandMapper {
-    int countByExample(PmsBrandExample example);
+    long countByExample(PmsBrandExample example);
 
     int deleteByExample(PmsBrandExample example);
